refactor(mjh/extraction): split jd_url_extractor.py (590 LOC) into fetcher + parser
- jd_url_fetcher.py — URL validation, auth-walled detection, httpx fetch,
  typed error hierarchy (JDFetchError / JDFetchTimeoutError / JDFetchAuthRequiredError)
- jd_url_parser.py — schema.org JobPosting fast path, HTML→text strip,
  Claude fallback, ExtractedJD dataclass
- jd_url_extractor.py — thin orchestrator + re-exports (no import-site changes
  required for existing callers)

Test patch paths updated to match the new module boundaries
(httpx patched on jd_url_fetcher, claude_service patched on jd_url_parser).
TECH_DEBT.md entry marked resolved.
